refactor(service-log): repoint optimistic add/edit/delete cache to the new query key and DTO shape
